+++ server/index.ts
@@ -1,5 +1,7 @@
-import express from 'express'
+import server from './server.js'
 
-const server = express()
+const PORT = process.env.PORT || 3000
 
-export default server
+server.listen(PORT, () => {
+  console.log(`Example app listening on port ${PORT}`)
+})
